Finish multi-threaded merge sort

+++ Student.h
@@ -5,7 +5,7 @@ typedef unsigned long long quint64;
 
 struct Student {
 	quint64 id;
-	float grade;
+	double grade;
 };
 
 struct ClassInfo {
